task 6

+++ 6-print_address.c
@@ -8,14 +8,14 @@
  */
 int print_address(va_list ap)
 {
-	unsigned long long int addr;
+	unsigned long int addr;
 	void *ptr = va_arg(ap, void *);
 
 	if (ptr == 0)
 	{
 		return (_printf("(null)"));
 	}
-	addr = (unsigned long long int)ptr;
+	addr = (unsigned long int)ptr;
 
 	return (_printf("0x%x", addr));
 }
